+++ codegen-compiler/src/main/kotlin/com/freeletics/khonshu/codegen/KhonshuSymbolProcessor.kt
@@ -1,8 +1,6 @@
 package com.freeletics.khonshu.codegen
 
 import com.freeletics.khonshu.codegen.codegen.FileGenerator
-import com.freeletics.khonshu.codegen.compose.NavDestination
-import com.freeletics.khonshu.codegen.compose.NavHostActivity
 import com.freeletics.khonshu.codegen.parser.toComposeScreenDestinationData
 import com.freeletics.khonshu.codegen.parser.toNavHostActivityData
 import com.google.auto.service.AutoService