+++ sample/simple/feature/main/src/main/kotlin/com/freeletics/khonshu/sample/feature/main/MainScreen.kt
@@ -4,8 +4,8 @@ import androidx.activity.ComponentActivity
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import com.freeletics.khonshu.codegen.compose.NavHostActivity
-import com.freeletics.khonshu.codegen.compose.SimpleNavHost
+import com.freeletics.khonshu.codegen.NavHostActivity
+import com.freeletics.khonshu.codegen.SimpleNavHost
 import com.freeletics.khonshu.sample.feature.root.nav.RootRoute
 
 @NavHostActivity(